+++ NEWS.md
@@ -1,6 +1,8 @@
 # qcbsRworkshops 0.1.0.9000
 
-* Simplify message when HTML presentations are successfully rendered.
+* Use GitHub Actions instead of Travis
+* `build_workshop()` gains an argument `script` to export R code.
+* Simplify messages displayed when HTML presentations are successfully rendered.
 * `install_workshops_pkgs_remotes()` use package `remotes` to install packages required for a given workshop.
 * `get_badges()` now returns 5 badges.
 * `get_badges()` gains argument `clip` to copy badges to the clipboard.